search bug fix

+++ src/main/java/nvt/kts/ticketapp/domain/dto/event/EventsDTO.java
@@ -11,20 +11,11 @@ import java.util.List;
 @NoArgsConstructor
 public class EventsDTO {
 
-//    private List<Event> events = new ArrayList<>();
-//
-//    private long numberOfElements;
-//
-//    public EventsDTO(List<Event> eventList, long numberOfElements) {
-//        this.events.addAll(eventList);
-//        this.numberOfElements = numberOfElements;
-//    }
-
     private List<EventDTO> events = new ArrayList<>();
 
-    private int numberOfElements;
+    private long numberOfElements;
 
-    public EventsDTO(List<EventDTO> events, int numberOfElements) {
+    public EventsDTO(List<EventDTO> events, long numberOfElements) {
         this.events = events;
         this.numberOfElements = numberOfElements;
     }
